+++ CHANGELOG.md
@@ -3,7 +3,8 @@ https://keepachangelog.com/en/1.0.0/
 
 ## [1.2.4] - 2022-08-31
 - Fixed: `AsyncOverloadsAvailable` supports methods that return `ValueTask`
-- 
+- Fixed: `AccessingTaskResultWithoutAwait` supports methods that return `ValueTask`
+- Fixed: 
 
 ## [1.2.3] - 2022-08-29
 - Fixed: `GetHashCodeRefersToMutableMember` correctly handles `partial` classes